Support for greedy matches for routes. This is required for odata support
Basically the term .** or /** matches everything in the rest of the url, therefore, should always be the last. A RoutingException is thrown if it's not (since the parsing term is pstring "**" .>> eol) it must be followed by end of line

The match goes to a route parameter conventially named "GreedyMatch".